Broke app down into separate apps.
Split models, views and admins.
General, misc. models and views in the main app.
