+++ umonya/apps/announcements/tests/__init__.py
@@ -1,0 +1,2 @@
+from test_page_rendering import *
+from test_home_page import *
