+++ umonya/apps/contact/models.py
@@ -1,0 +1,5 @@
+from django.db import models
+
+
+class Contact(models.Model):
+    pass
