+++ umonya/apps/contact/tests/__init__.py
@@ -1,0 +1,3 @@
+from test_page_rendering import *
+from test_email import *
+from test_forms import *
