+++ umonya/apps/course/tests/__init__.py
@@ -1,0 +1,1 @@
+from test_page_rendering import *
