+++ umonya/apps/main/templatetags/base_extras.py
@@ -1,15 +1,12 @@
 from django import template
-from django.core import urlresolvers
+from django.core.urlresolvers import reverse
 
 register = template.Library()
 
+
 @register.simple_tag
-def navactive(request, views):
-    views = ["umonya.apps.main.views." + view for view in views.split()]
+def navactive(request, urls):
     try:
-        view = urlresolvers.resolve(request.path).url_name
-        return view in views and "active" or ""
-    except urlresolvers.Resolver404:
-        return ""
-    except AttributeError:
+        return (request.path in (reverse(url) for url in urls.split())) and "active" or ""
+    except:
         return ""
