+++ umonya/apps/registration/admin.py
@@ -1,6 +1,7 @@
 from django.contrib import admin
-from models import Dynamic_Section
+from models import (Registration)
 
 # Ensuring that admin is aware of this class
 
-admin.site.register(Dynamic_Section)
+
+admin.site.register(Registration)
