+++ umonya/apps/registration/context_processors.py
@@ -1,6 +1,6 @@
 def dynamic_menu(request):
     # Adds Dynamic Menu Template Context
-    from models import Dynamic_Section
+    from apps.main.models import Dynamic_Section
     try:
         return{"registration": Dynamic_Section.objects.get(section="registration")}
     except:
