+++ umonya/apps/registration/tests/test_forms.py
@@ -1,13 +1,15 @@
-from umonya.apps.main.forms import RegistrationForm, ContactForm
+from apps.registration.forms import RegistrationForm
 from django.test import TestCase
-from umonya.apps.main.models import Registration, Dynamic_Section
-from umonya.apps.main.views import send_email_f
+from apps.main.models import Dynamic_Section
+from apps.registration.models import Registration
+from apps.main.views import send_email_f
 
 
 class TestForms(TestCase):
     """
         The tests below are failing, commenting them out for now
     """
+    pass
     # def test_form_rendering(self):
     #     Dynamic_Section.objects.create(section="registration", enabled=True)
     #     Registration.objects.create(name="name", field_type="CharField", text="What is Your Name?", priority=1, required=True)
@@ -25,28 +27,3 @@ class TestForms(TestCase):
     #     form = RegistrationForm(data=data)
     #     self.assertEqual(form["name"].errors, [u'This field is required.'])
     #     self.assertEqual(form["age"].errors, [u'This field is required.'])
-
-    def test_contactform_good_fields(self):
-        form_data = {"name": "Mark Gituma", "email": "Email@mail.com", "text": "Hello World"}
-        form = ContactForm(data=form_data)
-        self.assertIn("name", form.data)
-        self.assertIn("email", form.data)
-        self.assertIn("text", form.data)
-        if form.is_valid():
-            self.assertEqual(send_email_f(form), True)
-
-    def test_contactform_empty_fields(self):
-        form_data = {"name": "", "email": "", "text": ""}
-        form = ContactForm(data=form_data)
-        self.assertEqual(form["name"].errors, [u'This field is required.'])
-        self.assertEqual(form["email"].errors, [u'This field is required.'])
-
-    def test_contactform_bad_fields(self):
-        form_data = {"name": "", "email": "umonya", "text": ""}
-        form = ContactForm(data=form_data)
-        self.assertEqual(form["email"].errors, [u'Enter a valid email address.'])
-
-    def test_contact_form_post(self):
-        form_data = {"name": "Mark Gituma", "email": "Email@mail.com", "text": "Hello World"}
-        response = self.client.post("/contact/", form_data)
-        self.assertIn("success", response.context)
