+++ umonya/templates/navigation.html
@@ -15,7 +15,7 @@
       <div class="nav-collapse collapse" id="main-menu">
         <ul class="nav" id="main-menu-left">
           <li class="">
-            <a class="{% navactive request 'home view_announcement' %}" href="/">Home</a>
+            <a class="{% navactive request 'home' %}" href="/">Home</a>
           </li>
           <li class="">
             <a class="{% navactive request 'blog' %}" href="/blog">Blog</a>
